modulos/stock/requirements.txt
ffastapi
uvicorn[standard]
psycopg2-binary
kafka-python
python-jose[cryptography]
httpx
python-dotenv